integrate logger (not writing yet on the file)
