+++ projeto-intermedio/aeolus/src/main/java/lei/tqs/aeolus/externalAPI/APIResponse.java
@@ -1,4 +1,4 @@
-package lei.tqs.aeolus.ExternalAPI;
+package lei.tqs.aeolus.externalAPI;
 
 import lombok.*;
