+++ projeto-intermedio/aeolus/src/main/java/lei/tqs/aeolus/externalAPI/OpenWeatherUtils/Coordinate.java
@@ -1,4 +1,4 @@
-package lei.tqs.aeolus.ExternalAPI.OpenWeatherUtils;
+package lei.tqs.aeolus.externalAPI.OpenWeatherUtils;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.*;
