+++ projeto-intermedio/aeolus/src/main/resources/application.properties
@@ -1,9 +1,9 @@
 
 
 # logging
-logging.path=${APPDATA}/Logs
-logging.level.org.springframework=DEBUG
-logging.level.com.howtodoinjava=DEBUG
+logging.file.path=./logs/
+#logging.file.name=./logs/debug.log
+logging.level.lei.tqs.aeolus=DEBUG
 
 logging.pattern.console= %d{yyyy-MM-dd HH:mm:ss} %-5level - %msg%n
 logging.pattern.file= %d{yyyy-MM-dd HH:mm:ss} [%thread] %-5level %logger{36} - %msg%
